refactor e2e test
